Eu cred ca am facut cam toate functiile posibile pt repository. Verificati si voi si corelati cu cerinta sa vedeti daca am sarit ceva. Urmeaza implementare lor, numai selecturi adica, pe care o sa le faca teo, am vb cu ea acum

+++ src/Server/Model/RoomFare.java
@@ -2,13 +2,13 @@ package Server.Model;
 
 public class RoomFare {
     private String roomType;
-    private int pricePerDay;
+    private float pricePerDay;
 
     public String getRoomType() { return roomType; }
 
     public void setRoomType(String roomType) { this.roomType = roomType; }
 
-    public int getPricePerDay() { return pricePerDay; }
+    public float getPricePerDay() { return pricePerDay; }
 
     public void setPricePerDay(int pricePerDay) { this.pricePerDay = pricePerDay; }
 }
